+++ src/main/java/com/example/SpringLibrary/repository/BookRepository.java
@@ -3,10 +3,5 @@ package com.example.SpringLibrary.repository;
 import com.example.SpringLibrary.entity.Book;
 import org.springframework.data.repository.CrudRepository;
 
-
-// This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
-// CRUD refers Create, Read, Update, Delete
-
-public interface BookRepository extends CrudRepository<Book, Integer> {
-
+public interface BookRepository extends CrudRepository<Book, Long> {
 }
