+++ src/main/java/com/example/SpringLibrary/repository/UserRepository.java
@@ -1,8 +1,7 @@
 package com.example.SpringLibrary.repository;
 
-import com.example.SpringLibrary.entity.Book;
 import com.example.SpringLibrary.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 }
